Migrate composite filters
